+++ scripts/spells/lesser-restoration.js
@@ -1,7 +1,7 @@
 /*
 	You touch a creature and can end either one disease or one condition afflicting it. The condition can be Blinded, Deafened, Paralyzed, or Poisoned.
 */
-const version = "12.3.0";
+const version = "12.3.1";
 const optionName = "Lesser Restoration";
 const condition_list = ["Blinded", "Deafened", "Paralyzed", "Diseased", "Poisoned"];
 
@@ -9,59 +9,49 @@ try {
 	if (args[0].macroPass === "postActiveEffects") {
 		const targetToken = workflow.targets.first();
 		if (targetToken) {
-			const effect = targetToken.actor.effects.filter(i => condition_list.includes(i.name));
-			const selectOptions = effect.reduce((list, activeE) => {
-				let condition = activeE.label;
-				list.push(`<option value="${condition}">${condition}</option>`);
-				return list;
-			}, []);
+			const matchingEffects = HomebrewEffects.filterEffectsByConditions(targetToken.actor, condition_list);
 
-			if (selectOptions.length === 0)
+			let removedCondition;
+			if (matchingEffects.length === 0) {
 				return ui.notifications.error(`Nothing happens.. There's nothing to Cure on ${targetToken.actor.name}.`);
+			}
+			else if (matchingEffects.length === 1) {
+				removedCondition = matchingEffects[0];
+			}
+			else if (matchingEffects.length > 1) {
+				let target_list = await matchingEffects.reduce((list, target) => list += `<option value="${target.id}">${target.name}</option>`, ``);
 
-			let removedCondition = effect[0].label;
-
-			if (selectOptions.length > 1) {
-				let content = `
-					<div class="form-group">
-						<label><p>Which condition do you want to cure?</p></label>
-						<p>
-							<select name="element">${selectOptions.join('')}</select>
-						</p>
-					</div>`;
-
-				let dialog = new Promise((resolve, reject) => {
-					new Dialog({
-						title: `${optionName}`,
-						content,
-						buttons:
-							{
-								pick: {
-									label: `Remove it!`,
-									callback: async (html) => {
-										let cond = html.find('[name=element]')[0].value;
-										resolve(cond);
-									}
-								}
-							}
-					}).render(true);
+				removedCondition = await foundry.applications.api.DialogV2.prompt({
+					content: `<p>Which condition do you want to cure?</p><form><div class="form-group"><select id="condition">${target_list}</select></div></form>`,
+					rejectClose: false,
+					ok: {
+						callback: (event, button, dialog) => {
+							const effectId = button.form.elements.condition.value;
+							return targetToken.actor.getRollData().effects.find(e => e.id === effectId);
+						}
+					},
+					window: {
+						title: `${optionName}`
+					},
+					position: {
+						width: 400
+					}
 				});
 
-				removedCondition = await dialog;
 			}
 
-			let chosen = targetToken.actor.effects.find(i => i.name === removedCondition);
-			await MidiQOL.socket().executeAsGM("removeEffects", {
-				actorUuid: targetToken.actor.uuid,
-				effects: [chosen.id]
-			});
+			if (removedCondition) {
+				await MidiQOL.socket().executeAsGM("removeEffects", {
+					actorUuid: targetToken.actor.uuid,
+					effects: [removedCondition.id]
+				});
 
-			let chatContent = `<div class="midi-qol-nobox"><div class="midi-qol-flex-container"><div>Cures ${removedCondition}:</div><div class="midi-qol-target-npc midi-qol-target-name" id="${targetToken.id}"> ${targetToken.actor.name}</div><div><img src="${targetToken.document.texture.src}" width="30" height="30" style="border:0px"></img></div></div></div>`;
-
-			ChatMessage.create({
-				content: chatContent,
-				speaker: ChatMessage.getSpeaker({actor: actor})
-			});
+				let chatContent = `<div class="midi-qol-nobox"><div class="midi-qol-flex-container"><div>Cures ${removedCondition}:</div><div class="midi-qol-target-npc midi-qol-target-name" id="${targetToken.id}"> ${targetToken.actor.name}</div><div><img src="${targetToken.document.texture.src}" width="30" height="30" style="border:0px"></img></div></div></div>`;
+				ChatMessage.create({
+					content: chatContent,
+					speaker: ChatMessage.getSpeaker({actor: actor})
+				});
+			}
 		}
 	}